+++ python/ray/rllib/policy/rnn_sequencing.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
+
 """LSTM support for RLlib.
 
 The main trick here is that we add the time dimension at the last moment.
@@ -12,113 +13,11 @@ reshaping is possible.
 Note that this padding strategy only works out if we assume zero inputs don't
 meaningfully affect the loss function. This happens to be true for all the
 current algorithms: https://github.com/ray-project/ray/issues/2992
-
-See the add_time_dimension() and chop_into_sequences() functions below for
-more info.
 """
 
 import numpy as np
 
-from ray.rllib.models.tf.misc import linear, normc_initializer
-from ray.rllib.models.model import Model
-from ray.rllib.utils.annotations import override, DeveloperAPI, PublicAPI
-from ray.rllib.utils import try_import_tf
-
-tf = try_import_tf()
-
-
-# Deprecated: see as an alternative models/tf/recurrent_tf_modelv2.py
-class LSTM(Model):
-    """Adds a LSTM cell on top of some other model output.
-
-    Uses a linear layer at the end for output.
-
-    Important: we assume inputs is a padded batch of sequences denoted by
-        self.seq_lens. See add_time_dimension() for more information.
-    """
-
-    @override(Model)
-    def _build_layers_v2(self, input_dict, num_outputs, options):
-        cell_size = options.get("lstm_cell_size")
-        if options.get("lstm_use_prev_action_reward"):
-            action_dim = int(
-                np.product(
-                    input_dict["prev_actions"].get_shape().as_list()[1:]))
-            features = tf.concat(
-                [
-                    input_dict["obs"],
-                    tf.reshape(
-                        tf.cast(input_dict["prev_actions"], tf.float32),
-                        [-1, action_dim]),
-                    tf.reshape(input_dict["prev_rewards"], [-1, 1]),
-                ],
-                axis=1)
-        else:
-            features = input_dict["obs"]
-        last_layer = add_time_dimension(features, self.seq_lens)
-
-        # Setup the LSTM cell
-        lstm = tf.nn.rnn_cell.LSTMCell(cell_size, state_is_tuple=True)
-        self.state_init = [
-            np.zeros(lstm.state_size.c, np.float32),
-            np.zeros(lstm.state_size.h, np.float32)
-        ]
-
-        # Setup LSTM inputs
-        if self.state_in:
-            c_in, h_in = self.state_in
-        else:
-            c_in = tf.placeholder(
-                tf.float32, [None, lstm.state_size.c], name="c")
-            h_in = tf.placeholder(
-                tf.float32, [None, lstm.state_size.h], name="h")
-            self.state_in = [c_in, h_in]
-
-        # Setup LSTM outputs
-        state_in = tf.nn.rnn_cell.LSTMStateTuple(c_in, h_in)
-        lstm_out, lstm_state = tf.nn.dynamic_rnn(
-            lstm,
-            last_layer,
-            initial_state=state_in,
-            sequence_length=self.seq_lens,
-            time_major=False,
-            dtype=tf.float32)
-
-        self.state_out = list(lstm_state)
-
-        # Compute outputs
-        last_layer = tf.reshape(lstm_out, [-1, cell_size])
-        logits = linear(last_layer, num_outputs, "action",
-                        normc_initializer(0.01))
-        return logits, last_layer
-
-
-@PublicAPI
-def add_time_dimension(padded_inputs, seq_lens):
-    """Adds a time dimension to padded inputs.
-
-    Arguments:
-        padded_inputs (Tensor): a padded batch of sequences. That is,
-            for seq_lens=[1, 2, 2], then inputs=[A, *, B, B, C, C], where
-            A, B, C are sequence elements and * denotes padding.
-        seq_lens (Tensor): the sequence lengths within the input batch,
-            suitable for passing to tf.nn.dynamic_rnn().
-
-    Returns:
-        Reshaped tensor of shape [NUM_SEQUENCES, MAX_SEQ_LEN, ...].
-    """
-
-    # Sequence lengths have to be specified for LSTM batch inputs. The
-    # input batch must be padded to the max seq length given here. That is,
-    # batch_size == len(seq_lens) * max(seq_lens)
-    padded_batch_size = tf.shape(padded_inputs)[0]
-    max_seq_len = padded_batch_size // tf.shape(seq_lens)[0]
-
-    # Dynamically reshape the padded batch to introduce a time dimension.
-    new_batch_size = padded_batch_size // max_seq_len
-    new_shape = ([new_batch_size, max_seq_len] +
-                 padded_inputs.get_shape().as_list()[1:])
-    return tf.reshape(padded_inputs, new_shape)
+from ray.rllib.utils.annotations import DeveloperAPI
 
 
 @DeveloperAPI
